feat: add support for weighted products with decimal quantities
- Updated Product model to include `is_weighted` and `unit_label` fields.
- Modified stock and sale services to handle weighted products and decimal quantities.
- Enhanced API responses and tests to accommodate new product attributes.
- Updated frontend components to manage weighted product inputs and display.
- Added validation to prevent weighted products from having bundles.
- Implemented formatting for displaying quantities based on weight.

+++ sale/models.py
@@ -21,7 +21,7 @@ class Sale(BaseModel):
 class SaleLine(BaseModel):
     sale = models.ForeignKey(Sale, on_delete=models.CASCADE, related_name="lines")
     product = models.ForeignKey("product.Product", on_delete=models.PROTECT, related_name="+")
-    qty = models.PositiveIntegerField()
+    qty = models.DecimalField(max_digits=12, decimal_places=2)
     unit_price = models.IntegerField()
     line_total = models.IntegerField()
     is_bundle = models.BooleanField(default=False)
